and more refact
