+++ sources/tokenizer/variables_expansion.c
@@ -6,36 +6,11 @@
 /*   By: phemsi-a <phemsi-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/29 18:25:26 by phemsi-a          #+#    #+#             */
-/*   Updated: 2021/07/30 17:32:07 by phemsi-a         ###   ########.fr       */
+/*   Updated: 2021/07/30 17:46:22 by phemsi-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-static void	free_auxiliary_strings(char **temp, char **aux, char **splited_str)
-{
-	ft_free_and_null((void **)temp);
-	ft_free_and_null((void **)aux);
-	free_2d_array(splited_str);
-}
-
-static void	handle_quote(t_var *var, char **value, int i)
-{
-	char	**splited_string;
-	char	*temp;
-	char	*aux;
-
-	splited_string = ft_split(var->value, SINGLE_QUOTE);
-	ft_free_and_null((void **)&var->value);
-	var->value = ft_strdup(splited_string[0]);
-	if (splited_string[1])
-		temp = ft_strjoin("\'", splited_string[1]);
-	else
-		temp = ft_strdup("\'");
-	aux = ft_substr(*value, i, (ft_strlen(*value) - i));
-	var->after = ft_strjoin(temp, aux);
-	free_auxiliary_strings(&temp, &aux, splited_string);
-}
 
 static bool	is_exit_status_variable(char **value)
 {
@@ -53,10 +28,20 @@ static void	deal_with_string_slices(t_var *var, char **value, int *i)
 	var->before = ft_substr(*value, 0, *i);
 	var->value = ft_substr(*value, *i, get_var_size(var->pointer, i));
 	if (ft_strchr(var->value, SINGLE_QUOTE))
-		handle_quote(var, value, *i);
+		handle_var_single_quote(var, value, *i);
 	else
 		var->after = ft_substr(*value, *i, (ft_strlen(*value) - *i));
 	expand(&var->value);
+}
+
+static bool	has_dollar_sign(char **value, int *i, t_var *var)
+{
+	if (!(*value) || is_exit_status_variable(value))
+		return (FALSE);
+	var->pointer = search_var(*value, i);
+	if (!(var->pointer))
+		return (FALSE);
+	return (TRUE);
 }
 
 void	expand_variables(char **value)
@@ -66,10 +51,7 @@ void	expand_variables(char **value)
 	int		i;
 
 	i = 0;
-	if (!(*value) || is_exit_status_variable(value))
-		return ;
-	var.pointer = search_var(*value, &i);
-	if (!(var.pointer))
+	if (!has_dollar_sign(value, &i, &var))
 		return ;
 	deal_with_string_slices(&var, value, &i);
 	new_str = variadic_strjoin(3, var.before, var.value, var.after);
